multiagentapp/requirements.txt
google-adk==1.0.0
litellm==1.70.2
python-dotenv==1.1.0